Help liblo to detect Windows
Summary:
liblo needs preprocessor symbol `WIN32` defined in order to
detect that it is compiled on Windows.  MinGW compilers do
not have `WIN32` defined by default, but `_WIN32` is defined.
This does **not** need to be modified for 64 bit Windows.
`_WIN32` is the correct macro to check for the Windows
platform, regardless of 32 or 64 bit operating system.

Test Plan: Code review.

Reviewers: giso

Reviewed By: giso

Differential Revision: https://dev.openmha.org/D639