Replace random() with rand(), unify usages
Summary:
random() is a POSIX function and not available on Windows.  In order to
prepare the Windows port, random() must be replaced.  This change
replaces random() with rand(), which is good enough to generate noise
in audio signal processing, and unifies usages of our drand() helper
function.

Test Plan: Test Builds on Linux (PC and ARM) and Mac were successful.

Reviewers: giso

Reviewed By: giso

Tags: #tascar

Differential Revision: https://dev.openmha.org/D634