Compile TASCAR main GUI for Windows (not for production yet)
Summary:
* Change Jenkinsfile to:
** compile tascar.exe for Windows.
** compile tascar_hrirconv.dll for Windows.
** make build artifacts better available from Jenkins.
** add a Jenkinsfile.differential which can be used in future
   to compile and test changes on Linux, ARM and Windows before
   committing them to git.
* Extend README.windows with additional build dependency for
  tascar.exe and update compilation commands to compile
  tascar.exe and hrirconv.dll.
* Fix GUI Makefile: Tweak dependencies to ensure
  spkcalib_glade.h has been created when it is needed.
* Add #ifdefs to tascar_mainwindow.cc for windows compilation
  with reduced functionality (no webkit2 on Windows).

Test Plan:
Checked compilation and execution of compiled tascar.exe.  With
these changes, tascar.exe can be compiled and started for
windows.
It generates audio and shows movements on the map.  There are
some problems when the scene file contains non-ASCII characters,
but when using only ASCII then everything seems to work as long
as the CPU load does not cause dropouts.

Reviewers: giso

Reviewed By: giso

Subscribers: #tascar

Tags: #tascar

Differential Revision: https://dev.openmha.org/D655