Move pipeline job part to dedicated pipeline agent
The previous pipeline any declaration on the overall agent
declaration of the Jenkins pipeline could potentially block
agents with scarce resources.

A new Jenkins agent is available for these pipeline jobs
which do nothing but evaluate the Jenkinsfile and then wait
on other stages executing on other agents doing the real work,
with label "pipeline".  It has 10 execution slots but only 2
CPUs, therefore it should only be used in places like these
changes where most of the time is spent in waiting.